requirements.txt
python-dotenv==0.21.0
jsonschema==4.17.3
requests==2.28.1
python-i18n==0.3.9
questionary==1.10.0
google-genai==1.0.0
